Partial port of Kilostation
It's playable TECHNICALLY but beware, NO SHUTTLES SPAWN YET so if you join after round-start you're just gonna be stuck in space 'til one of us figures it out, I'm going to sleep now so try to help me figure out why no shuttles spawn. They've been added to the datums and the docking ports in the map are assigned correctly, think it might just be a "me" issue, also this does not add Kilo to the rotation yet, so it can't be selected from a vote.